+++ src/goose/__init__.py
@@ -2,4 +2,4 @@ from .pluginbase.notifier import Notifier  # noqa: F401
 from .pluginbase.toolkit import Toolkit, tool  # noqa: F401
 from .pluginbase.profile import Profile, ToolkitSpec  # noqa: F401
 from .pluginbase.command import Command  # noqa: F401
-from .pluginbase import utils  # noqa: F401
+from .pluginbase.utils import converter  # noqa: F401
