+++ src/goose/_internal/toolkit/repo_context/repo_context_utils.py
@@ -9,7 +9,7 @@ from exchange import Exchange
 from goose.pluginbase.utils.ask import ask_an_ai
 
 
-def get_directory_size(directory: str) -> int:
+def _get_directory_size(directory: str) -> int:
     total_size = 0
     for dirpath, _, filenames in os.walk(directory):
         for f in filenames:
@@ -23,10 +23,10 @@ def get_directory_size(directory: str) -> int:
 def get_repo_size(repo_path: str) -> int:
     """Returns repo size in MB"""
     git_dir = os.path.join(repo_path, ".git")
-    return get_directory_size(git_dir) / (1024**2)
+    return _get_directory_size(git_dir) / (1024**2)
 
 
-def get_files_and_directories(root_dir: str) -> Dict[str, list]:
+def _get_files_and_directories(root_dir: str) -> Dict[str, list]:
     """Gets file names and directory names. Checks that goose has correctly typed the file and directory names and that
     the files actually exist (to avoid downstream file read errors).
 
@@ -70,7 +70,7 @@ def goose_picks_files(root: str, exchange: Exchange, max_workers: int = 4) -> Li
     with concurrent.futures.ThreadPoolExecutor(max_workers=max_workers) as executor:
         while queue:
             current_batch = [queue.popleft() for _ in range(min(max_workers, len(queue)))]
-            futures = {executor.submit(process_directory, dir, exchange): dir for dir in current_batch}
+            futures = {executor.submit(_process_directory, dir, exchange): dir for dir in current_batch}
 
             for future in concurrent.futures.as_completed(futures):
                 files, next_dirs = future.result()
@@ -80,12 +80,12 @@ def goose_picks_files(root: str, exchange: Exchange, max_workers: int = 4) -> Li
     return all_files
 
 
-def process_directory(current_dir: str, exchange: Exchange) -> Tuple[List[str], List[str]]:
+def _process_directory(current_dir: str, exchange: Exchange) -> Tuple[List[str], List[str]]:
     """Allows goose to pick files and subdirectories contained in a given directory (current_dir). Get the list of file
     and directory names in the current folder, then ask Goose to pick which ones to keep.
 
     """
-    files_and_dirs = get_files_and_directories(current_dir)
+    files_and_dirs = _get_files_and_directories(current_dir)
     ai_response = ask_an_ai(str(files_and_dirs), exchange)
 
     # FIXME: goose response validation
