+++ tests/_internal/test_utils.py
@@ -1,0 +1,7 @@
+from goose._internal.utils import load_plugins
+
+
+def test_load_plugins():
+    plugins = load_plugins("exchange.provider")
+    assert isinstance(plugins, dict)
+    assert len(plugins) > 0
