+++ tests/pluginbase/utils/test_converter.py
@@ -1,7 +1,5 @@
-import string
-
 import pytest
-from goose.pluginbase.utils import droid, ensure, ensure_list, load_plugins
+from goose.pluginbase.utils.converter import ensure, ensure_list
 
 
 class MockClass:
@@ -10,13 +8,6 @@ class MockClass:
 
     def __eq__(self, other):
         return self.name == other.name
-
-
-def test_load_plugins():
-    plugins = load_plugins("exchange.provider")
-    assert isinstance(plugins, dict)
-    assert len(plugins) > 0
-
 
 def test_ensure_with_class():
     mock_class = MockClass("foo")
@@ -52,12 +43,3 @@ def test_ensure_list():
     obj_list = ensure_list(MockClass)(["foo", "bar"])
     assert obj_list == [MockClass("foo"), MockClass("bar")]
 
-
-def test_droid():
-    result = droid()
-    assert isinstance(result, str)
-    assert len(result) == 4
-    for character in [result[i] for i in [0, 2]]:
-        assert character in string.ascii_lowercase, "should be in lower case"
-    for character in [result[i] for i in [1, 3]]:
-        assert character in string.digits, "should be a digit"
